Updated styling for removing inactive mentors from mentors' list (#226)
Updated styling for inactive mentors to a custom style. This avoids re-adding them to the list as empty cards when the "Clear filter" button is clicked.

When 'Clear filter' button is clicked, the current logic removes the display class (.d-none) from all mentor cards and that included inactive mentors. This causes the cards for inactive mentors to show but is empty.